fix(indexer): repair agent-arena indexer compilation and local stack
- Upgrade tokio-tungstenite to 0.28 and adapt API changes (Message::Text/Ping signatures)
- Fix solana_subscriber connect_async call and notification field access
- Add missing Solana config fields to indexer Config
- Fix mappers and db queries (TaskListSort variants, dynamic ORDER BY, i32 state)
- Remove duplicate wrapper functions in main.rs and clean unused imports
- Add Dockerfile builder deps (pkg-config, libssl-dev) for release build
- Fix frontend build: replace next/dynamic with client mount guard in Providers.tsx
- Add cross-service integration test (agent-daemon) covering auth, profile, feed APIs
- Update workspace dependency references for local development compatibility

+++ apps/agent-arena/indexer/src/mappers.rs
@@ -30,7 +30,7 @@ pub fn map_task(task: crate::db::TaskRow) -> TaskApi {
 }
 
 /// Map task state integer to string
-fn map_task_state(state: i16) -> String {
+fn map_task_state(state: i32) -> String {
     match state {
         0 => "open".to_string(),
         1 => "completed".to_string(),
